+++ conf/aggregator.prod.py
@@ -18,13 +18,25 @@ DEFAULT_OIDC_CLIENT_EGI = {
 configured_oidc_providers = [
     OidcProvider(
         id="egi",
-        issuer="https://aai.egi.eu/oidc/",
+        issuer="https://aai.egi.eu/auth/realms/egi/",
         scopes=[
             "openid", "email",
             "eduperson_entitlement",
             "eduperson_scoped_affiliation",
         ],
         title="EGI Check-in",
+        default_client=DEFAULT_OIDC_CLIENT_EGI,  # TODO: remove this legacy experimental field
+        default_clients=[DEFAULT_OIDC_CLIENT_EGI],
+    ),
+    OidcProvider(
+        id="egi-old",
+        issuer="https://aai.egi.eu/oidc/",  # TODO: remove old EGI provider refs (issuer https://aai.egi.eu/oidc/)
+        scopes=[
+            "openid", "email",
+            "eduperson_entitlement",
+            "eduperson_scoped_affiliation",
+        ],
+        title="EGI Check-in (legacy)",
         default_client=DEFAULT_OIDC_CLIENT_EGI,  # TODO: remove this legacy experimental field
         default_clients=[DEFAULT_OIDC_CLIENT_EGI],
     ),
@@ -36,7 +48,10 @@ config = AggregatorConfig(
         "vito": "https://openeo.vito.be/openeo/1.0/",
         "eodc": "https://openeo.eodc.eu/v1.0/",
     },
-    auth_entitlement_check={"oidc_issuer_whitelist": {"https://aai.egi.eu/oidc"}},
+    auth_entitlement_check={"oidc_issuer_whitelist": {
+        "https://aai.egi.eu/auth/realms/egi/",
+        "https://aai.egi.eu/oidc",  # TODO: remove old EGI provider refs (issuer https://aai.egi.eu/oidc/)
+    }},
     configured_oidc_providers=configured_oidc_providers,
     partitioned_job_tracking={
         "zk_hosts": "epod-master1.vgt.vito.be:2181,epod-master2.vgt.vito.be:2181,epod-master3.vgt.vito.be:2181",